% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{create_activation}
\alias{create_activation}
\title{Registers your on-premises server or virtual machine with Amazon EC2 so that you can manage these resources using Run Command}
\usage{
create_activation(Description = NULL, DefaultInstanceName = NULL,
  IamRole, RegistrationLimit = NULL, ExpirationDate = NULL)
}
\arguments{
\item{Description}{A user-defined description of the resource that you want to register with Amazon EC2.

Do not enter personally identifiable information in this field.}

\item{DefaultInstanceName}{The name of the registered, managed instance as it will appear in the Amazon EC2 console or when you use the AWS command line tools to list EC2 resources.

Do not enter personally identifiable information in this field.}

\item{IamRole}{[required] The Amazon Identity and Access Management (IAM) role that you want to assign to the managed instance.}

\item{RegistrationLimit}{Specify the maximum number of managed instances you want to register. The default value is 1 instance.}

\item{ExpirationDate}{The date by which this activation request should expire. The default value is 24 hours.}
}
\description{
Registers your on-premises server or virtual machine with Amazon EC2 so that you can manage these resources using Run Command. An on-premises server or virtual machine that has been registered with EC2 is called a managed instance. For more information about activations, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/systems-manager-managedinstances.html}{Setting Up Systems Manager in Hybrid Environments}.
}
\section{Accepted Parameters}{
\preformatted{create_activation(
  Description = "string",
  DefaultInstanceName = "string",
  IamRole = "string",
  RegistrationLimit = 123,
  ExpirationDate = as.POSIXct("2015-01-01")
)
}
}

