% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{create_association}
\alias{create_association}
\title{Associates the specified Systems Manager document with the specified instances or targets}
\usage{
create_association(Name, DocumentVersion = NULL, InstanceId = NULL,
  Parameters = NULL, Targets = NULL, ScheduleExpression = NULL,
  OutputLocation = NULL, AssociationName = NULL)
}
\arguments{
\item{Name}{The name of the Systems Manager document.}

\item{DocumentVersion}{The document version you want to associate with the target(s). Can be a specific version or the default version.}

\item{InstanceId}{The instance ID.}

\item{Parameters}{The parameters for the documents runtime configuration.}

\item{Targets}{The targets (either instances or tags) for the association.}

\item{ScheduleExpression}{A cron expression when the association will be applied to the target(s).}

\item{OutputLocation}{An Amazon S3 bucket where you want to store the output details of the request.}

\item{AssociationName}{Specify a descriptive name for the association.}
}
\description{
Associates the specified Systems Manager document with the specified instances or targets.
}
\details{
When you associate a document with one or more instances using instance IDs or tags, SSM Agent running on the instance processes the document and configures the instance as specified.

If you associate a document with an instance that already has an associated document, the system throws the AssociationAlreadyExists exception.
}
\examples{

}
