% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{create_resource_data_sync}
\alias{create_resource_data_sync}
\title{Creates a resource data sync configuration to a single bucket in Amazon S3}
\usage{
create_resource_data_sync(
  SyncName = "string",
  S3Destination = list(
    BucketName = "string",
    Prefix = "string",
    SyncFormat = "JsonSerDe",
    Region = "string",
    AWSKMSKeyARN = "string"
  )
)
}
\arguments{
\item{SyncName}{[required] A name for the configuration.}

\item{S3Destination}{[required] Amazon S3 configuration details for the sync.}
}
\description{
Creates a resource data sync configuration to a single bucket in Amazon S3. This is an asynchronous operation that returns immediately. After a successful initial sync is completed, the system continuously syncs data to the Amazon S3 bucket. To check the status of the sync, use the ListResourceDataSync.
}
\details{
By default, data is not encrypted in Amazon S3. We strongly recommend that you enable encryption in Amazon S3 to ensure secure data storage. We also recommend that you secure access to the Amazon S3 bucket by creating a restrictive bucket policy. To view an example of a restrictive Amazon S3 bucket policy for Resource Data Sync, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/sysman-inventory-datasync-create.html}{Create a Resource Data Sync for Inventory} in the \emph{AWS Systems Manager User Guide}.
}
\examples{

}
