% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{delete_activation}
\alias{delete_activation}
\title{Deletes an activation}
\usage{
delete_activation(ActivationId)
}
\arguments{
\item{ActivationId}{The ID of the activation that you want to delete.}
}
\description{
Deletes an activation. You are not required to delete an activation. If you delete an activation, you can no longer use it to register additional managed instances. Deleting an activation does not de-register managed instances. You must manually de-register managed instances.
}
\examples{

}
