% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{delete_maintenance_window}
\alias{delete_maintenance_window}
\title{Deletes a Maintenance Window}
\usage{
delete_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window to delete.}
}
\description{
Deletes a Maintenance Window.
}
\section{Accepted Parameters}{
\preformatted{delete_maintenance_window(
  WindowId = "string"
)
}
}

