% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{delete_parameter}
\alias{delete_parameter}
\title{Delete a parameter from the system}
\usage{
delete_parameter(Name)
}
\arguments{
\item{Name}{The name of the parameter to delete.}
}
\description{
Delete a parameter from the system.
}
\examples{

}
