% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{delete_parameters}
\alias{delete_parameters}
\title{Delete a list of parameters}
\usage{
delete_parameters(Names)
}
\arguments{
\item{Names}{[required] The names of the parameters to delete.}
}
\description{
Delete a list of parameters. This API is used to delete parameters by using the Amazon EC2 console.
}
\section{Accepted Parameters}{
\preformatted{delete_parameters(
  Names = list(
    "string"
  )
)
}
}

