% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{delete_patch_baseline}
\alias{delete_patch_baseline}
\title{Deletes a patch baseline}
\usage{
delete_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to delete.}
}
\description{
Deletes a patch baseline.
}
\section{Accepted Parameters}{
\preformatted{delete_patch_baseline(
  BaselineId = "string"
)
}
}

