% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{deregister_patch_baseline_for_patch_group}
\alias{deregister_patch_baseline_for_patch_group}
\title{Removes a patch group from a patch baseline}
\usage{
deregister_patch_baseline_for_patch_group(
  BaselineId = "string",
  PatchGroup = "string"
)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to deregister the patch group from.}

\item{PatchGroup}{[required] The name of the patch group that should be deregistered from the patch baseline.}
}
\description{
Removes a patch group from a patch baseline.
}
\examples{

}
