% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{deregister_task_from_maintenance_window}
\alias{deregister_task_from_maintenance_window}
\title{Removes a task from a Maintenance Window}
\usage{
deregister_task_from_maintenance_window(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window the task should be removed from.}

\item{WindowTaskId}{[required] The ID of the task to remove from the Maintenance Window.}
}
\description{
Removes a task from a Maintenance Window.
}
\section{Accepted Parameters}{
\preformatted{deregister_task_from_maintenance_window(
  WindowId = "string",
  WindowTaskId = "string"
)
}
}

