% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_association_execution_targets}
\alias{describe_association_execution_targets}
\title{Use this API action to view information about a specific execution of a specific association}
\usage{
describe_association_execution_targets(AssociationId, ExecutionId,
  Filters = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{AssociationId}{[required] The association ID that includes the execution for which you want to view details.}

\item{ExecutionId}{[required] The execution ID for which you want to view details.}

\item{Filters}{Filters for the request. You can specify the following filters and values.

Status (EQUAL)

ResourceId (EQUAL)

ResourceType (EQUAL)}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of results.}
}
\description{
Use this API action to view information about a specific execution of a specific association.
}
\section{Accepted Parameters}{
\preformatted{describe_association_execution_targets(
  AssociationId = "string",
  ExecutionId = "string",
  Filters = list(
    list(
      Key = "Status"|"ResourceId"|"ResourceType",
      Value = "string"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

