% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_available_patches}
\alias{describe_available_patches}
\title{Lists all patches that could possibly be included in a patch baseline}
\usage{
describe_available_patches(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{Filters}{Filters used to scope down the returned patches.}

\item{MaxResults}{The maximum number of patches to return (per page).}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists all patches that could possibly be included in a patch baseline.
}
\examples{

}
