% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_document}
\alias{describe_document}
\title{Describes the specified Systems Manager document}
\usage{
describe_document(Name, DocumentVersion = NULL)
}
\arguments{
\item{Name}{[required] The name of the Systems Manager document.}

\item{DocumentVersion}{The document version for which you want information. Can be a specific version or the default version.}
}
\description{
Describes the specified Systems Manager document.
}
\section{Accepted Parameters}{
\preformatted{describe_document(
  Name = "string",
  DocumentVersion = "string"
)
}
}

