% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_effective_instance_associations}
\alias{describe_effective_instance_associations}
\title{All associations for the instance(s)}
\usage{
describe_effective_instance_associations(
  InstanceId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{InstanceId}{[required] The instance ID for which you want to view all associations.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
All associations for the instance(s).
}
\examples{

}
