% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_effective_patches_for_patch_baseline}
\alias{describe_effective_patches_for_patch_baseline}
\title{Retrieves the current effective patches (the patch and the approval state) for the specified patch baseline}
\usage{
describe_effective_patches_for_patch_baseline(
  BaselineId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to retrieve the effective patches for.}

\item{MaxResults}{The maximum number of patches to return (per page).}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Retrieves the current effective patches (the patch and the approval state) for the specified patch baseline. Note that this API applies only to Windows patch baselines.
}
\examples{

}
