% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_instance_patches}
\alias{describe_instance_patches}
\title{Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance}
\usage{
describe_instance_patches(InstanceId, Filters = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{InstanceId}{The ID of the instance whose patch state information should be retrieved.}

\item{Filters}{Each entry in the array is a structure containing:

Key (string, between 1 and 128 characters)

Values (array of strings, each string between 1 and 256 characters)}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}

\item{MaxResults}{The maximum number of patches to return (per page).}
}
\description{
Retrieves information about the patches on the specified instance and their state relative to the patch baseline being used for the instance.
}
\examples{

}
