% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_inventory_deletions}
\alias{describe_inventory_deletions}
\title{Describes a specific delete inventory operation}
\usage{
describe_inventory_deletions(DeletionId = NULL, NextToken = NULL,
  MaxResults = NULL)
}
\arguments{
\item{DeletionId}{Specify the delete inventory ID for which you want information. This ID was returned by the \code{DeleteInventory} action.}

\item{NextToken}{A token to start the list. Use this token to get the next set of results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}
}
\description{
Describes a specific delete inventory operation.
}
\examples{

}
