% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_maintenance_window_execution_task_invocations}
\alias{describe_maintenance_window_execution_task_invocations}
\title{Retrieves the individual task executions (one per target) for a particular task executed as part of a Maintenance Window execution}
\usage{
describe_maintenance_window_execution_task_invocations(WindowExecutionId,
  TaskId, Filters = NULL, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{WindowExecutionId}{The ID of the Maintenance Window execution the task is part of.}

\item{TaskId}{The ID of the specific task in the Maintenance Window task that should be retrieved.}

\item{Filters}{Optional filters used to scope down the returned task invocations. The supported filter key is STATUS with the corresponding values PENDING, IN\_PROGRESS, SUCCESS, FAILED, TIMED\_OUT, CANCELLING, and CANCELLED.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Retrieves the individual task executions (one per target) for a particular task executed as part of a Maintenance Window execution.
}
\examples{

}
