% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_maintenance_window_execution_tasks}
\alias{describe_maintenance_window_execution_tasks}
\title{For a given Maintenance Window execution, lists the tasks that were executed}
\usage{
describe_maintenance_window_execution_tasks(
  WindowExecutionId = "string",
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution whose task executions should be retrieved.}

\item{Filters}{Optional filters used to scope down the returned tasks. The supported filter key is STATUS with the corresponding values PENDING, IN\_PROGRESS, SUCCESS, FAILED, TIMED\_OUT, CANCELLING, and CANCELLED.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
For a given Maintenance Window execution, lists the tasks that were executed.
}
\examples{

}
