% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_maintenance_window_targets}
\alias{describe_maintenance_window_targets}
\title{Lists the targets registered with the Maintenance Window}
\usage{
describe_maintenance_window_targets(WindowId, Filters = NULL,
  MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window whose targets should be retrieved.}

\item{Filters}{Optional filters that can be used to narrow down the scope of the returned window targets. The supported filter keys are Type, WindowTargetId and OwnerInformation.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists the targets registered with the Maintenance Window.
}
\section{Accepted Parameters}{
\preformatted{describe_maintenance_window_targets(
  WindowId = "string",
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

