% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_maintenance_window_tasks}
\alias{describe_maintenance_window_tasks}
\title{Lists the tasks in a Maintenance Window}
\usage{
describe_maintenance_window_tasks(WindowId, Filters = NULL,
  MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window whose tasks should be retrieved.}

\item{Filters}{Optional filters used to narrow down the scope of the returned tasks. The supported filter keys are WindowTaskId, TaskArn, Priority, and TaskType.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists the tasks in a Maintenance Window.
}
\section{Accepted Parameters}{
\preformatted{describe_maintenance_window_tasks(
  WindowId = "string",
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

