% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_maintenance_windows}
\alias{describe_maintenance_windows}
\title{Retrieves the Maintenance Windows in an AWS account}
\usage{
describe_maintenance_windows(Filters = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{Filters}{Optional filters used to narrow down the scope of the returned Maintenance Windows. Supported filter keys are \strong{Name} and \strong{Enabled}.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Retrieves the Maintenance Windows in an AWS account.
}
\section{Accepted Parameters}{
\preformatted{describe_maintenance_windows(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

