% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_patch_baselines}
\alias{describe_patch_baselines}
\title{Lists the patch baselines in your AWS account}
\usage{
describe_patch_baselines(Filters = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{Filters}{Each element in the array is a structure containing:

Key: (string, "NAME\_PREFIX" or "OWNER")

Value: (array of strings, exactly 1 entry, between 1 and 255 characters)}

\item{MaxResults}{The maximum number of patch baselines to return (per page).}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists the patch baselines in your AWS account.
}
\examples{

}
