% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_patch_group_state}
\alias{describe_patch_group_state}
\title{Returns high-level aggregated patch compliance state for a patch group}
\usage{
describe_patch_group_state(PatchGroup)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch snapshot should be retrieved.}
}
\description{
Returns high-level aggregated patch compliance state for a patch group.
}
\section{Accepted Parameters}{
\preformatted{describe_patch_group_state(
  PatchGroup = "string"
)
}
}

