% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{describe_patch_groups}
\alias{describe_patch_groups}
\title{Lists all patch groups that have been registered with patch baselines}
\usage{
describe_patch_groups(MaxResults = NULL, Filters = NULL, NextToken = NULL)
}
\arguments{
\item{MaxResults}{The maximum number of patch groups to return (per page).}

\item{Filters}{One or more filters. Use a filter to return a more specific list of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists all patch groups that have been registered with patch baselines.
}
\examples{

}
