% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_automation_execution}
\alias{get_automation_execution}
\title{Get detailed information about a particular Automation execution}
\usage{
get_automation_execution(AutomationExecutionId)
}
\arguments{
\item{AutomationExecutionId}{[required] The unique identifier for an existing automation execution to examine. The execution ID is returned by StartAutomationExecution when the execution of an Automation document is initiated.}
}
\description{
Get detailed information about a particular Automation execution.
}
\section{Accepted Parameters}{
\preformatted{get_automation_execution(
  AutomationExecutionId = "string"
)
}
}

