% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_command_invocation}
\alias{get_command_invocation}
\title{Returns detailed information about command execution for an invocation or plugin}
\usage{
get_command_invocation(
  CommandId = "string",
  InstanceId = "string",
  PluginName = "string"
)
}
\arguments{
\item{CommandId}{[required] (Required) The parent command ID of the invocation plugin.}

\item{InstanceId}{[required] (Required) The ID of the managed instance targeted by the command. A managed instance can be an Amazon EC2 instance or an instance in your hybrid environment that is configured for Systems Manager.}

\item{PluginName}{(Optional) The name of the plugin for which you want detailed results. If the document contains only one plugin, the name can be omitted and the details will be returned.}
}
\description{
Returns detailed information about command execution for an invocation or plugin.
}
\examples{

}
