% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_connection_status}
\alias{get_connection_status}
\title{Retrieves the Session Manager connection status for an instance to determine whether it is connected and ready to receive Session Manager connections}
\usage{
get_connection_status(Target)
}
\arguments{
\item{Target}{[required] The ID of the instance.}
}
\description{
Retrieves the Session Manager connection status for an instance to determine whether it is connected and ready to receive Session Manager connections.
}
\section{Accepted Parameters}{
\preformatted{get_connection_status(
  Target = "string"
)
}
}

