% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_default_patch_baseline}
\alias{get_default_patch_baseline}
\title{Retrieves the default patch baseline}
\usage{
get_default_patch_baseline(OperatingSystem = NULL)
}
\arguments{
\item{OperatingSystem}{Returns the default patch baseline for the specified operating system.}
}
\description{
Retrieves the default patch baseline. Note that Systems Manager supports creating multiple default patch baselines. For example, you can create a default patch baseline for each operating system.
}
\details{
If you do not specify an operating system value, the default patch baseline for Windows is returned.
}
\section{Accepted Parameters}{
\preformatted{get_default_patch_baseline(
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"
)
}
}

