% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_document}
\alias{get_document}
\title{Gets the contents of the specified Systems Manager document}
\usage{
get_document(Name, DocumentVersion = NULL, DocumentFormat = NULL)
}
\arguments{
\item{Name}{[required] The name of the Systems Manager document.}

\item{DocumentVersion}{The document version for which you want information.}

\item{DocumentFormat}{Returns the document in the specified format. The document format can be either JSON or YAML. JSON is the default format.}
}
\description{
Gets the contents of the specified Systems Manager document.
}
\section{Accepted Parameters}{
\preformatted{get_document(
  Name = "string",
  DocumentVersion = "string",
  DocumentFormat = "YAML"|"JSON"
)
}
}

