% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_maintenance_window}
\alias{get_maintenance_window}
\title{Retrieves a Maintenance Window}
\usage{
get_maintenance_window(WindowId)
}
\arguments{
\item{WindowId}{[required] The ID of the desired Maintenance Window.}
}
\description{
Retrieves a Maintenance Window.
}
\section{Accepted Parameters}{
\preformatted{get_maintenance_window(
  WindowId = "string"
)
}
}

