% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_maintenance_window_execution}
\alias{get_maintenance_window_execution}
\title{Retrieves details about a specific task executed as part of a Maintenance Window execution}
\usage{
get_maintenance_window_execution(
  WindowExecutionId = "string"
)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution that includes the task.}
}
\description{
Retrieves details about a specific task executed as part of a Maintenance Window execution.
}
\examples{

}
