% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_maintenance_window_execution_task}
\alias{get_maintenance_window_execution_task}
\title{Retrieves the details about a specific task executed as part of a Maintenance Window execution}
\usage{
get_maintenance_window_execution_task(WindowExecutionId, TaskId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution that includes the task.}

\item{TaskId}{[required] The ID of the specific task execution in the Maintenance Window task that should be retrieved.}
}
\description{
Retrieves the details about a specific task executed as part of a Maintenance Window execution.
}
\section{Accepted Parameters}{
\preformatted{get_maintenance_window_execution_task(
  WindowExecutionId = "string",
  TaskId = "string"
)
}
}

