% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_maintenance_window_execution_task_invocation}
\alias{get_maintenance_window_execution_task_invocation}
\title{Retrieves a task invocation}
\usage{
get_maintenance_window_execution_task_invocation(WindowExecutionId, TaskId,
  InvocationId)
}
\arguments{
\item{WindowExecutionId}{[required] The ID of the Maintenance Window execution for which the task is a part.}

\item{TaskId}{[required] The ID of the specific task in the Maintenance Window task that should be retrieved.}

\item{InvocationId}{[required] The invocation ID to retrieve.}
}
\description{
Retrieves a task invocation. A task invocation is a specific task executing on a specific target. Maintenance Windows report status for all invocations.
}
\section{Accepted Parameters}{
\preformatted{get_maintenance_window_execution_task_invocation(
  WindowExecutionId = "string",
  TaskId = "string",
  InvocationId = "string"
)
}
}

