% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_maintenance_window_task}
\alias{get_maintenance_window_task}
\title{Lists the tasks in a Maintenance Window}
\usage{
get_maintenance_window_task(WindowId, WindowTaskId)
}
\arguments{
\item{WindowId}{The Maintenance Window ID that includes the task to retrieve.}

\item{WindowTaskId}{The Maintenance Window task ID to retrieve.}
}
\description{
Lists the tasks in a Maintenance Window.
}
\examples{

}
