% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_parameter}
\alias{get_parameter}
\title{Get information about a parameter by using the parameter name}
\usage{
get_parameter(Name, WithDecryption = NULL)
}
\arguments{
\item{Name}{[required] The name of the parameter you want to query.}

\item{WithDecryption}{Return decrypted values for secure string parameters. This flag is ignored for String and StringList parameter types.}
}
\description{
Get information about a parameter by using the parameter name. Don't confuse this API action with the GetParameters API action.
}
\section{Accepted Parameters}{
\preformatted{get_parameter(
  Name = "string",
  WithDecryption = TRUE|FALSE
)
}
}

