% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_parameter_history}
\alias{get_parameter_history}
\title{Query a list of all parameters used by the AWS account}
\usage{
get_parameter_history(Name, WithDecryption = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{Name}{The name of a parameter you want to query.}

\item{WithDecryption}{Return decrypted values for secure string parameters. This flag is ignored for String and StringList parameter types.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Query a list of all parameters used by the AWS account.
}
\examples{

}
