% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_parameters}
\alias{get_parameters}
\title{Get details of a parameter}
\usage{
get_parameters(Names, WithDecryption = NULL)
}
\arguments{
\item{Names}{Names of the parameters for which you want to query information.}

\item{WithDecryption}{Return decrypted secure string value. Return decrypted values for secure string parameters. This flag is ignored for String and StringList parameter types.}
}
\description{
Get details of a parameter. Don't confuse this API action with the GetParameter API action.
}
\examples{

}
