% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_patch_baseline}
\alias{get_patch_baseline}
\title{Retrieves information about a patch baseline}
\usage{
get_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to retrieve.}
}
\description{
Retrieves information about a patch baseline.
}
\section{Accepted Parameters}{
\preformatted{get_patch_baseline(
  BaselineId = "string"
)
}
}

