% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{list_association_versions}
\alias{list_association_versions}
\title{Retrieves all versions of an association for a specific association ID}
\usage{
list_association_versions(AssociationId, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{AssociationId}{[required] The association ID for which you want to view all versions.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of results.}
}
\description{
Retrieves all versions of an association for a specific association ID.
}
\section{Accepted Parameters}{
\preformatted{list_association_versions(
  AssociationId = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

