% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{list_associations}
\alias{list_associations}
\title{Lists the associations for the specified Systems Manager document or instance}
\usage{
list_associations(AssociationFilterList = NULL, MaxResults = NULL,
  NextToken = NULL)
}
\arguments{
\item{AssociationFilterList}{One or more filters. Use a filter to return a more specific list of results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
Lists the associations for the specified Systems Manager document or instance.
}
\section{Accepted Parameters}{
\preformatted{list_associations(
  AssociationFilterList = list(
    list(
      key = "InstanceId"|"Name"|"AssociationId"|"AssociationStatusName"|"LastExecutedBefore"|"LastExecutedAfter"|"AssociationName",
      value = "string"
    )
  ),
  MaxResults = 123,
  NextToken = "string"
)
}
}

