% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{list_commands}
\alias{list_commands}
\title{Lists the commands requested by users of the AWS account}
\usage{
list_commands(CommandId = NULL, InstanceId = NULL, MaxResults = NULL,
  NextToken = NULL, Filters = NULL)
}
\arguments{
\item{CommandId}{(Optional) If provided, lists only the specified command.}

\item{InstanceId}{(Optional) Lists commands issued against this instance ID.}

\item{MaxResults}{(Optional) The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{(Optional) The token for the next set of items to return. (You received this token from a previous call.)}

\item{Filters}{(Optional) One or more filters. Use a filter to return a more specific list of results.}
}
\description{
Lists the commands requested by users of the AWS account.
}
\section{Accepted Parameters}{
\preformatted{list_commands(
  CommandId = "string",
  InstanceId = "string",
  MaxResults = 123,
  NextToken = "string",
  Filters = list(
    list(
      key = "InvokedAfter"|"InvokedBefore"|"Status"|"ExecutionStage"|"DocumentName",
      value = "string"
    )
  )
)
}
}

