% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{list_document_versions}
\alias{list_document_versions}
\title{List all versions for a document}
\usage{
list_document_versions(Name, MaxResults = NULL, NextToken = NULL)
}
\arguments{
\item{Name}{[required] The name of the document about which you want version information.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}

\item{NextToken}{The token for the next set of items to return. (You received this token from a previous call.)}
}
\description{
List all versions for a document.
}
\section{Accepted Parameters}{
\preformatted{list_document_versions(
  Name = "string",
  MaxResults = 123,
  NextToken = "string"
)
}
}

