% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{list_resource_compliance_summaries}
\alias{list_resource_compliance_summaries}
\title{Returns a resource-level summary count}
\usage{
list_resource_compliance_summaries(
  Filters = list(
    list(
      Key = "string",
      Values = list(
        "string"
      ),
      Type = "EQUAL"|"NOT_EQUAL"|"BEGIN_WITH"|"LESS_THAN"|"GREATER_THAN"
    )
  ),
  NextToken = "string",
  MaxResults = 123
)
}
\arguments{
\item{Filters}{One or more filters. Use a filter to return a more specific list of results.}

\item{NextToken}{A token to start the list. Use this token to get the next set of results.}

\item{MaxResults}{The maximum number of items to return for this call. The call also returns a token that you can specify in a subsequent call to get the next set of results.}
}
\description{
Returns a resource-level summary count. The summary includes information about compliant and non-compliant statuses and detailed compliance-item severity counts, according to the filter criteria you specify.
}
\examples{

}
