% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{register_default_patch_baseline}
\alias{register_default_patch_baseline}
\title{Defines the default patch baseline}
\usage{
register_default_patch_baseline(BaselineId)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline that should be the default patch baseline.}
}
\description{
Defines the default patch baseline.
}
\section{Accepted Parameters}{
\preformatted{register_default_patch_baseline(
  BaselineId = "string"
)
}
}

