% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{resume_session}
\alias{resume_session}
\title{Reconnects a session to an instance after it has been disconnected}
\usage{
resume_session(
  SessionId = "string"
)
}
\arguments{
\item{SessionId}{[required] The ID of the disconnected session to resume.}
}
\description{
Reconnects a session to an instance after it has been disconnected. Connections can be resumed for disconnected sessions, but not terminated sessions.
}
\details{
This command is primarily for use by client machines to automatically reconnect during intermittent network issues. It is not intended for any other use.
}
\examples{

}
