% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{send_command}
\alias{send_command}
\title{Executes commands on one or more managed instances}
\usage{
send_command(
  InstanceIds = list(
    "string"
  ),
  Targets = list(
    list(
      Key = "string",
      Values = list(
        "string"
      )
    )
  ),
  DocumentName = "string",
  DocumentVersion = "string",
  DocumentHash = "string",
  DocumentHashType = "Sha256"|"Sha1",
  TimeoutSeconds = 123,
  Comment = "string",
  Parameters = list(
    list(
      "string"
    )
  ),
  OutputS3Region = "string",
  OutputS3BucketName = "string",
  OutputS3KeyPrefix = "string",
  MaxConcurrency = "string",
  MaxErrors = "string",
  ServiceRoleArn = "string",
  NotificationConfig = list(
    NotificationArn = "string",
    NotificationEvents = list(
      "All"|"InProgress"|"Success"|"TimedOut"|"Cancelled"|"Failed"
    ),
    NotificationType = "Command"|"Invocation"
  ),
  CloudWatchOutputConfig = list(
    CloudWatchLogGroupName = "string",
    CloudWatchOutputEnabled = TRUE|FALSE
  )
)
}
\arguments{
\item{InstanceIds}{The instance IDs where the command should execute. You can specify a maximum of 50 IDs. If you prefer not to list individual instance IDs, you can instead send commands to a fleet of instances using the Targets parameter, which accepts EC2 tags. For more information about how to use targets, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/send-commands-multiple.html}{Sending Commands to a Fleet} in the \emph{AWS Systems Manager User Guide}.}

\item{Targets}{(Optional) An array of search criteria that targets instances using a Key,Value combination that you specify. Targets is required if you don't provide one or more instance IDs in the call. For more information about how to use targets, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/send-commands-multiple.html}{Sending Commands to a Fleet} in the \emph{AWS Systems Manager User Guide}.}

\item{DocumentName}{[required] Required. The name of the Systems Manager document to execute. This can be a public document or a custom document.}

\item{DocumentVersion}{The SSM document version to use in the request. You can specify \$DEFAULT, \$LATEST, or a specific version number. If you execute commands by using the AWS CLI, then you must escape the first two options by using a backslash. If you specify a version number, then you don't need to use the backslash. For example:

--document-version "\\$DEFAULT"

--document-version "\\$LATEST"

--document-version "3"}

\item{DocumentHash}{The Sha256 or Sha1 hash created by the system when the document was created.

Sha1 hashes have been deprecated.}

\item{DocumentHashType}{Sha256 or Sha1.

Sha1 hashes have been deprecated.}

\item{TimeoutSeconds}{If this time is reached and the command has not already started executing, it will not run.}

\item{Comment}{User-specified information about the command, such as a brief description of what the command should do.}

\item{Parameters}{The required and optional parameters specified in the document being executed.}

\item{OutputS3Region}{(Deprecated) You can no longer specify this parameter. The system ignores it. Instead, Systems Manager automatically determines the Amazon S3 bucket region.}

\item{OutputS3BucketName}{The name of the S3 bucket where command execution responses should be stored.}

\item{OutputS3KeyPrefix}{The directory structure within the S3 bucket where the responses should be stored.}

\item{MaxConcurrency}{(Optional) The maximum number of instances that are allowed to execute the command at the same time. You can specify a number such as 10 or a percentage such as 10%. The default value is 50. For more information about how to use MaxConcurrency, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/send-commands-multiple.html#send-commands-velocity}{Using Concurrency Controls} in the \emph{AWS Systems Manager User Guide}.}

\item{MaxErrors}{The maximum number of errors allowed without the command failing. When the command fails one more time beyond the value of MaxErrors, the systems stops sending the command to additional targets. You can specify a number like 10 or a percentage like 10%. The default value is 0. For more information about how to use MaxErrors, see \href{http://docs.aws.amazon.com/systems-manager/latest/userguide/send-commands-multiple.html#send-commands-maxerrors}{Using Error Controls} in the \emph{AWS Systems Manager User Guide}.}

\item{ServiceRoleArn}{The IAM role that Systems Manager uses to send notifications.}

\item{NotificationConfig}{Configurations for sending notifications.}

\item{CloudWatchOutputConfig}{Enables Systems Manager to send Run Command output to Amazon CloudWatch Logs.}
}
\description{
Executes commands on one or more managed instances.
}
\examples{

}
