% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{start_associations_once}
\alias{start_associations_once}
\title{Use this API action to execute an association immediately and only one time}
\usage{
start_associations_once(AssociationIds)
}
\arguments{
\item{AssociationIds}{[required] The association IDs that you want to execute immediately and only one time.}
}
\description{
Use this API action to execute an association immediately and only one time. This action can be helpful when troubleshooting associations.
}
\section{Accepted Parameters}{
\preformatted{start_associations_once(
  AssociationIds = list(
    "string"
  )
)
}
}

