% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{terminate_session}
\alias{terminate_session}
\title{Permanently ends a session and closes the data connection between the Session Manager client and SSM Agent on the instance}
\usage{
terminate_session(SessionId)
}
\arguments{
\item{SessionId}{The ID of the session to terminate.}
}
\description{
Permanently ends a session and closes the data connection between the Session Manager client and SSM Agent on the instance. A terminated session cannot be resumed.
}
\examples{

}
