% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_association}
\alias{update_association}
\title{Updates an association}
\usage{
update_association(AssociationId, Parameters = NULL, DocumentVersion = NULL,
  ScheduleExpression = NULL, OutputLocation = NULL, Name = NULL,
  Targets = NULL, AssociationName = NULL, AssociationVersion = NULL)
}
\arguments{
\item{AssociationId}{The ID of the association you want to update.}

\item{Parameters}{The parameters you want to update for the association. If you create a parameter using Parameter Store, you can reference the parameter using {{ssm:parameter-name}}}

\item{DocumentVersion}{The document version you want update for the association.}

\item{ScheduleExpression}{The cron expression used to schedule the association that you want to update.}

\item{OutputLocation}{An Amazon S3 bucket where you want to store the results of this request.}

\item{Name}{The name of the association document.}

\item{Targets}{The targets of the association.}

\item{AssociationName}{The name of the association that you want to update.}

\item{AssociationVersion}{This parameter is provided for concurrency control purposes. You must specify the latest association version in the service. If you want to ensure that this request succeeds, either specify \code{$LATEST}, or omit this parameter.}
}
\description{
Updates an association. You can update the association name and version, the document version, schedule, parameters, and Amazon S3 output.
}
\examples{

}
