% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_association_status}
\alias{update_association_status}
\title{Updates the status of the Systems Manager document associated with the specified instance}
\usage{
update_association_status(Name, InstanceId, AssociationStatus)
}
\arguments{
\item{Name}{[required] The name of the Systems Manager document.}

\item{InstanceId}{[required] The ID of the instance.}

\item{AssociationStatus}{[required] The association status.}
}
\description{
Updates the status of the Systems Manager document associated with the specified instance.
}
\section{Accepted Parameters}{
\preformatted{update_association_status(
  Name = "string",
  InstanceId = "string",
  AssociationStatus = list(
    Date = as.POSIXct("2015-01-01"),
    Name = "Pending"|"Success"|"Failed",
    Message = "string",
    AdditionalInfo = "string"
  )
)
}
}

