% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_document}
\alias{update_document}
\title{The document you want to update}
\usage{
update_document(Content, Name, DocumentVersion = NULL,
  DocumentFormat = NULL, TargetType = NULL)
}
\arguments{
\item{Content}{The content in a document that you want to update.}

\item{Name}{The name of the document that you want to update.}

\item{DocumentVersion}{The version of the document that you want to update.}

\item{DocumentFormat}{Specify the document format for the new document version. Systems Manager supports JSON and YAML documents. JSON is the default format.}

\item{TargetType}{Specify a new target type for the document.}
}
\description{
The document you want to update.
}
\examples{

}
