% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_maintenance_window}
\alias{update_maintenance_window}
\title{Updates an existing Maintenance Window}
\usage{
update_maintenance_window(
  WindowId = "string",
  Name = "string",
  Description = "string",
  StartDate = "string",
  EndDate = "string",
  Schedule = "string",
  ScheduleTimezone = "string",
  Duration = 123,
  Cutoff = 123,
  AllowUnassociatedTargets = TRUE|FALSE,
  Enabled = TRUE|FALSE,
  Replace = TRUE|FALSE
)
}
\arguments{
\item{WindowId}{[required] The ID of the Maintenance Window to update.}

\item{Name}{The name of the Maintenance Window.}

\item{Description}{An optional description for the update request.}

\item{StartDate}{The time zone that the scheduled Maintenance Window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los\_Angeles", "etc/UTC", or "Asia/Seoul". For more information, see the \href{https://www.iana.org/time-zones}{Time Zone Database} on the IANA website.}

\item{EndDate}{The date and time, in ISO-8601 Extended format, for when you want the Maintenance Window to become inactive. EndDate allows you to set a date and time in the future when the Maintenance Window will no longer run.}

\item{Schedule}{The schedule of the Maintenance Window in the form of a cron or rate expression.}

\item{ScheduleTimezone}{The time zone that the scheduled Maintenance Window executions are based on, in Internet Assigned Numbers Authority (IANA) format. For example: "America/Los\_Angeles", "etc/UTC", or "Asia/Seoul". For more information, see the \href{https://www.iana.org/time-zones}{Time Zone Database} on the IANA website.}

\item{Duration}{The duration of the Maintenance Window in hours.}

\item{Cutoff}{The number of hours before the end of the Maintenance Window that Systems Manager stops scheduling new tasks for execution.}

\item{AllowUnassociatedTargets}{Whether targets must be registered with the Maintenance Window before tasks can be defined for those targets.}

\item{Enabled}{Whether the Maintenance Window is enabled.}

\item{Replace}{If True, then all fields that are required by the CreateMaintenanceWindow action are also required for this API request. Optional fields that are not specified are set to null.}
}
\description{
Updates an existing Maintenance Window. Only specified parameters are modified.
}
\examples{

}
