% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_maintenance_window_target}
\alias{update_maintenance_window_target}
\title{Modifies the target of an existing Maintenance Window}
\usage{
update_maintenance_window_target(WindowId, WindowTargetId, Targets = NULL,
  OwnerInformation = NULL, Name = NULL, Description = NULL,
  Replace = NULL)
}
\arguments{
\item{WindowId}{The Maintenance Window ID with which to modify the target.}

\item{WindowTargetId}{The target ID to modify.}

\item{Targets}{The targets to add or replace.}

\item{OwnerInformation}{User-provided value that will be included in any CloudWatch events raised while running tasks for these targets in this Maintenance Window.}

\item{Name}{A name for the update.}

\item{Description}{An optional description for the update.}

\item{Replace}{If True, then all fields that are required by the RegisterTargetWithMaintenanceWindow action are also required for this API request. Optional fields that are not specified are set to null.}
}
\description{
Modifies the target of an existing Maintenance Window. You can't change the target type, but you can change the following:
}
\details{
The target from being an ID target to a Tag target, or a Tag target to an ID target.

IDs for an ID target.

Tags for a Tag target.

Owner.

Name.

Description.

If a parameter is null, then the corresponding field is not modified.
}
\examples{

}
