% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_deployable_patch_snapshot_for_instance}
\alias{get_deployable_patch_snapshot_for_instance}
\title{Retrieves the current snapshot for the patch baseline the instance uses}
\usage{
get_deployable_patch_snapshot_for_instance(InstanceId, SnapshotId)
}
\arguments{
\item{InstanceId}{The ID of the instance for which the appropriate patch snapshot should be retrieved.}

\item{SnapshotId}{The user-defined snapshot ID.}
}
\description{
Retrieves the current snapshot for the patch baseline the instance uses. This API is primarily used by the AWS-RunPatchBaseline Systems Manager document.
}
\examples{

}
