% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{get_patch_baseline_for_patch_group}
\alias{get_patch_baseline_for_patch_group}
\title{Retrieves the patch baseline that should be used for the specified patch group}
\usage{
get_patch_baseline_for_patch_group(PatchGroup, OperatingSystem = NULL)
}
\arguments{
\item{PatchGroup}{[required] The name of the patch group whose patch baseline should be retrieved.}

\item{OperatingSystem}{Returns he operating system rule specified for patch groups using the patch baseline.}
}
\description{
Retrieves the patch baseline that should be used for the specified patch group.
}
\section{Accepted Parameters}{
\preformatted{get_patch_baseline_for_patch_group(
  PatchGroup = "string",
  OperatingSystem = "WINDOWS"|"AMAZON_LINUX"|"AMAZON_LINUX_2"|"UBUNTU"|"REDHAT_ENTERPRISE_LINUX"|"SUSE"|"CENTOS"
)
}
}

