% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{register_patch_baseline_for_patch_group}
\alias{register_patch_baseline_for_patch_group}
\title{Registers a patch baseline for a patch group}
\usage{
register_patch_baseline_for_patch_group(BaselineId, PatchGroup)
}
\arguments{
\item{BaselineId}{[required] The ID of the patch baseline to register the patch group with.}

\item{PatchGroup}{[required] The name of the patch group that should be registered with the patch baseline.}
}
\description{
Registers a patch baseline for a patch group.
}
\section{Accepted Parameters}{
\preformatted{register_patch_baseline_for_patch_group(
  BaselineId = "string",
  PatchGroup = "string"
)
}
}

