% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_association_status}
\alias{update_association_status}
\title{Updates the status of the Systems Manager document associated with the specified instance}
\usage{
update_association_status(Name, InstanceId, AssociationStatus)
}
\arguments{
\item{Name}{The name of the Systems Manager document.}

\item{InstanceId}{The ID of the instance.}

\item{AssociationStatus}{The association status.}
}
\description{
Updates the status of the Systems Manager document associated with the specified instance.
}
\examples{

}
