% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_document_default_version}
\alias{update_document_default_version}
\title{Set the default version of a document}
\usage{
update_document_default_version(Name, DocumentVersion)
}
\arguments{
\item{Name}{The name of a custom document that you want to set as the default version.}

\item{DocumentVersion}{The version of a custom document that you want to set as the default version.}
}
\description{
Set the default version of a document.
}
\examples{

}
