% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.ssm_operations.R
\name{update_managed_instance_role}
\alias{update_managed_instance_role}
\title{Assigns or changes an Amazon Identity and Access Management (IAM) role to the managed instance}
\usage{
update_managed_instance_role(InstanceId, IamRole)
}
\arguments{
\item{InstanceId}{[required] The ID of the managed instance where you want to update the role.}

\item{IamRole}{[required] The IAM role you want to assign or change.}
}
\description{
Assigns or changes an Amazon Identity and Access Management (IAM) role to the managed instance.
}
\section{Accepted Parameters}{
\preformatted{update_managed_instance_role(
  InstanceId = "string",
  IamRole = "string"
)
}
}

